Moved logic from DeletePost.tsx into useDeletePost hook.
