+++ client/src/components/modal/DeletePost.tsx
@@ -1,9 +1,5 @@
 import React from "react";
-import { useAuth } from "../../contexts/AuthContext";
-import useAxiosPrivate from "../../hooks/useAxiosPrivate";
-import { toast } from "react-hot-toast";
-import { useTheme } from "../../contexts/ThemeContext";
-import { useNavigate } from "react-router-dom";
+import useDeletePost from "../../hooks/posts/useDeletePost";
 
 type Props = {
   postId: string
@@ -19,43 +15,8 @@ export default function DeletePost({
   deletePostById,
   type,
 }: Props) {
-  const { currentUser } = useAuth();
-  const axiosPrivate = useAxiosPrivate();
-  const { theme } = useTheme();
-  const navigate = useNavigate();
-  const endpoint =
-    type === "POST"
-      ? "/api/posts"
-      : type === "COMMENT"
-      ? "api/comments"
-      : undefined;
+  const { deletePost } = useDeletePost({ postId, parentId, onDeletePost: deletePostById, setShowModal, type });
 
-  const deletePost = async (postId: string) => {
-    try {
-      const response = await axiosPrivate.delete(`${endpoint}/${postId}`, {
-        data: { userId: currentUser._id, parentId: parentId },
-      });
-      deletePostById(response.data._id);
-      setShowModal(false);
-      toast.success("Post has been deleted!", {
-        style: {
-          backgroundColor: `${theme === "light" ? "" : "#16181c"}`,
-          color: `${theme === "light" ? "" : "#fff"}`,
-        },
-      });
-    } catch (err: any) {
-      console.error(err);
-      if (err.response?.status === 403) {
-        navigate("/login", { state: { from: location }, replace: true });
-      }
-      toast.error("Error!", {
-        style: {
-          backgroundColor: `${theme === "light" ? "" : "#16181c"}`,
-          color: `${theme === "light" ? "" : "#fff"}`,
-        },
-      });
-    }
-  };
   return (
     <div>
       <h1 className="modal-centered">Delete Post?</h1>
@@ -69,7 +30,7 @@ export default function DeletePost({
         >
           Cancel
         </button>
-        <button onClick={() => deletePost(postId)} className="delete-btn">
+        <button onClick={(deletePost)} className="delete-btn">
           Delete Post
         </button>{" "}
       </div>
